fix(project) - rename project components folder
